@@ -1,6 +1,6 @@
 {
   "name": "react-dom",
-  "version": "18.3.1",
+  "version": "19.2.4",
   "description": "React package for working with the DOM.",
   "main": "index.js",
   "repository": {
@@ -15,48 +15,103 @@
   "bugs": {
     "url": "https://github.com/facebook/react/issues"
   },
-  "homepage": "https://reactjs.org/",
+  "homepage": "https://react.dev/",
   "dependencies": {
-    "loose-envify": "^1.1.0",
-    "scheduler": "^0.23.2"
+    "scheduler": "^0.27.0"
   },
   "peerDependencies": {
-    "react": "^18.3.1"
+    "react": "^19.2.4"
   },
   "files": [
     "LICENSE",
     "README.md",
-    "index.js",
     "client.js",
+    "client.react-server.js",
+    "index.js",
     "profiling.js",
-    "server.js",
+    "profiling.react-server.js",
+    "react-dom.react-server.js",
     "server.browser.js",
+    "server.bun.js",
+    "server.edge.js",
+    "server.js",
     "server.node.js",
+    "server.react-server.js",
+    "static.browser.js",
+    "static.edge.js",
+    "static.js",
+    "static.node.js",
+    "static.react-server.js",
     "test-utils.js",
-    "cjs/",
-    "umd/"
+    "cjs/"
   ],
   "exports": {
-    ".": "./index.js",
-    "./client": "./client.js",
+    ".": {
+      "react-server": "./react-dom.react-server.js",
+      "default": "./index.js"
+    },
+    "./client": {
+      "react-server": "./client.react-server.js",
+      "default": "./client.js"
+    },
     "./server": {
+      "react-server": "./server.react-server.js",
+      "workerd": "./server.edge.js",
+      "bun": "./server.bun.js",
       "deno": "./server.browser.js",
       "worker": "./server.browser.js",
+      "node": "./server.node.js",
+      "edge-light": "./server.edge.js",
       "browser": "./server.browser.js",
       "default": "./server.node.js"
     },
-    "./server.browser": "./server.browser.js",
-    "./server.node": "./server.node.js",
-    "./profiling": "./profiling.js",
+    "./server.browser": {
+      "react-server": "./server.react-server.js",
+      "default": "./server.browser.js"
+    },
+    "./server.bun": {
+      "react-server": "./server.react-server.js",
+      "default": "./server.bun.js"
+    },
+    "./server.edge": {
+      "react-server": "./server.react-server.js",
+      "default": "./server.edge.js"
+    },
+    "./server.node": {
+      "react-server": "./server.react-server.js",
+      "default": "./server.node.js"
+    },
+    "./static": {
+      "react-server": "./static.react-server.js",
+      "workerd": "./static.edge.js",
+      "deno": "./static.browser.js",
+      "worker": "./static.browser.js",
+      "node": "./static.node.js",
+      "edge-light": "./static.edge.js",
+      "browser": "./static.browser.js",
+      "default": "./static.node.js"
+    },
+    "./static.browser": {
+      "react-server": "./static.react-server.js",
+      "default": "./static.browser.js"
+    },
+    "./static.edge": {
+      "react-server": "./static.react-server.js",
+      "default": "./static.edge.js"
+    },
+    "./static.node": {
+      "react-server": "./static.react-server.js",
+      "default": "./static.node.js"
+    },
+    "./profiling": {
+      "react-server": "./profiling.react-server.js",
+      "default": "./profiling.js"
+    },
     "./test-utils": "./test-utils.js",
     "./package.json": "./package.json"
   },
   "browser": {
-    "./server.js": "./server.browser.js"
-  },
-  "browserify": {
-    "transform": [
-      "loose-envify"
-    ]
+    "./server.js": "./server.browser.js",
+    "./static.js": "./static.browser.js"
   }
 }
