@@ -1,41 +1,33 @@
 {
-    "name": "krsnathkr-portfolio",
-    "private": true,
-    "version": "0.0.0",
-    "homepage": "https://krsnathkr.github.io",
-    "type": "module",
-    "scripts": {
-        "dev": "vite",
-        "build": "vite build",
-        "lint": "eslint . --ext js,jsx --report-unused-disable-directives --max-warnings 0",
-        "preview": "vite preview",
-        "predeploy": "npm run build",
-        "deploy": "gh-pages -d dist"
-    },
-    "dependencies": {
-        "clsx": "^2.1.1",
-        "framer-motion": "^10.16.4",
-        "gsap": "^3.13.0",
-        "lucide-react": "^0.292.0",
-        "mathjs": "^15.1.0",
-        "ogl": "^1.0.11",
-        "react": "^18.2.0",
-        "react-dom": "^18.2.0",
-        "tailwind-merge": "^3.4.0",
-        "three": "^0.181.2"
-    },
-    "devDependencies": {
-        "@types/react": "^18.2.37",
-        "@types/react-dom": "^18.2.15",
-        "@vitejs/plugin-react": "^4.2.0",
-        "autoprefixer": "^10.4.16",
-        "eslint": "^8.53.0",
-        "eslint-plugin-react": "^7.33.2",
-        "eslint-plugin-react-hooks": "^4.6.0",
-        "eslint-plugin-react-refresh": "^0.4.4",
-        "gh-pages": "^6.3.0",
-        "postcss": "^8.4.31",
-        "tailwindcss": "^3.3.5",
-        "vite": "^5.0.0"
-    }
+  "name": "krsnathkr.github.io",
+  "private": true,
+  "version": "0.0.0",
+  "type": "module",
+  "scripts": {
+    "dev": "vite",
+    "build": "vite build",
+    "lint": "eslint .",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "@tailwindcss/postcss": "^4.1.18",
+    "autoprefixer": "^10.4.24",
+    "framer-motion": "^12.34.0",
+    "lucide-react": "^0.564.0",
+    "postcss": "^8.5.6",
+    "react": "^19.2.0",
+    "react-dom": "^19.2.0",
+    "tailwindcss": "^4.1.18"
+  },
+  "devDependencies": {
+    "@eslint/js": "^9.39.1",
+    "@types/react": "^19.2.7",
+    "@types/react-dom": "^19.2.3",
+    "@vitejs/plugin-react": "^5.1.1",
+    "eslint": "^9.39.1",
+    "eslint-plugin-react-hooks": "^7.0.1",
+    "eslint-plugin-react-refresh": "^0.4.24",
+    "globals": "^16.5.0",
+    "vite": "^7.3.1"
+  }
 }
